requirements.txt
selenium==3.141.0
allure-pytest==2.8.7
pytest==5.3.4
